[bcs / type tag] Fix some bugs, add error handling, full testing (#28)
* [bcs] Add deserialization errors

This adds errors when deserializing a type isn't possible due to too few bytes

* [bcs] Add full test coverage and bugfixes for BCS package

* [typetag] Add documentation, missing types, and cut down extra data

There were some holder pieces of data in the TypeTags that weren't used
since TypeTags are just a representation of a type.  Those were removed
and tests added to properly cover to 89% of lines in the TypeTags.

Additionally, some helper functions were added for option, string and
object types.